+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Remove components `SplashErrorFatal` and `SplashErrorNetwork`
+
+- Update default icon component `SplashError`
+
+- Update styles component `NonIdealState`. Remove `isError` prop
+
 ## [0.39.1] - 2021-07-12
 
 - Update prop `handleTabChange` in `Tabbed` component;